functions/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
httpx>=0.27.0